LPS-54989 move packages
