+++ modules/apps/document-library/document-library-repository-search-impl/src/com/liferay/document/library/repository/search/util/KeywordsUtil.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.util.lucene;
+package com.liferay.document.library.repository.search.util;
 
 import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.kernel.util.StringUtil;
