+++ modules/apps/document-library/document-library-repository-search-impl/test/unit/com/liferay/document/library/repository/search/util/KeywordsUtilTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.util.lucene;
+package com.liferay.document.library.repository.search.util;
 
 import org.junit.Assert;
 import org.junit.Test;
